libcamerausb: resolve the dead lock of the still capture.
BZ: 120437

Issue:
When doing still capture, the shutter sound will be sent to upper.
So the ControlThread is blocked by this shutter sound callback, it just can keep on running after the shutter sound callback done.
But in the camera service, when it receive the shutter sound, it will get parameters from the camera HAL. But the get parameters will be blocked by the still capture command.
So the dead lock will happen.

Solution:
Porting the CallbacksThread from the libcamera2 to the libcamerausb. It can resolve the problem.
But we don't need move all the callbacks to the callbacks thread. We just put the shutter sound callback to the CallbacksThread right now.

Change-Id: I82128dfe8681ea21882219e143958ccbd3efaf79
Signed-off-by: Yang Liang <liang.l.yang@intel.com>
Reviewed-on: http://android.intel.com:8080/116467
Reviewed-by: Hu, Gang A <gang.a.hu@intel.com>
Tested-by: Hu, Gang A <gang.a.hu@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>