libcamerausb: 720p video recording is not stable
BZ: 121529

Issue:
The 720p video recording can work, but it will fail to do recording sometimes.
In the Camera HAL, when the storeMetaDataInVideoBuffers is called, it needs to
restart the preview to allocate buffers. And the mState will be changed.
If the restart preview has been done before the previewEnabled is called, the recording
will success.
If the restart preview doesn't finish and the previewEnabled has been called,
the recording will fail.
Because the mState will be changed when doing the restart preview.

Solution:
add one mutex lock to make sure that the mState is after the restart preview

Change-Id: Iae4f79a77fc3a54c517a575a17ff321fc299afde
Signed-off-by: Yang Liang <liang.l.yang@intel.com>
Reviewed-on: http://android.intel.com:8080/117357
Reviewed-by: Hu, Gang A <gang.a.hu@intel.com>
Tested-by: Hu, Gang A <gang.a.hu@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>