[Camera Convergence] Fix and enable snapshot-during-video capability
This was a pre-existing-but-apparently-bitrotten capability that had
been disabled.  Fix a typo that returned incorrect video frame sizes
from CameraDriver.  Refactor thumbnail handling and enable it for
snapshots as the encoding logic now apparently requires it (and it's
very fast: there's no reason not to have it enabled even for recording
operations, especially on a multicore CPU with idle threads).

Change-Id: I92f3cacfab33f45517c4a6e0b8f73f3de84bb290
Signed-off-by: Andy Ross <andy.ross@windriver.com>
Signed-off-by: Sreenidhi Koti <sreenidhi.koti@intel.com>