[PORT FROM JB] libcamerausb: Fix CTS testPreviewPictureSizesCombination failure
BZ: 159713

Issue:
1: when take picture, the preview won't stop, so when change preview size,
the callback data is wrong for application, so need stop preview when doing capture
2: when take picture, it will stop preview, return buffer
(just send returnBuffer msg and waiting for the takePicture msg been done).
and it will release jpegdec buffer. allocate new buffer for capture,
but the handle may be the same with released jpegdec buffer,such as A.
when the returnBuffer is processed, it shouldn't be returned, because all the
buffer had been released, but buffer A will be returned(capture type), so encoder failed.

Solution:
1: for the first issue, stop preview when take picture
2: for the second issue, return capture buffer in stopCapture

Change-Id: I15ccafc6588bd7aede7c17aa79d747984272201f
Signed-off-by: jingzhi yu <jingzhi.yu@intel.com>
Reviewed-on: https://android.intel.com/175378
Reviewed-by: Sripathi, Srinivas <srinivas.sripathi@intel.com>
Tested-by: sys_pccgand <sys_pccgand@intel.com>
Reviewed-by: Chittim Munirathnam, MadhusudanX <madhusudanx.chittim.munirathnam@intel.com>
Tested-by: Chittim Munirathnam, MadhusudanX <madhusudanx.chittim.munirathnam@intel.com>