From 833344a81768b11f107a89da84fbd4a4c81bf725 Mon Sep 17 00:00:00 2001 From: Chon Ming Lee <chon.ming.lee@intel.com> Date: Thu, 25 Sep 2014 22:54:52 +0800 Subject: [PATCH] libcamerausb: Fix no camera attached causing mediaserver crash
The libcamerausb always assume camera is attached.  If no camera attached, no
camera status doesn't pass back to the camera service.  As a result, camera
service assume the camera is there, and continue to execute, as a result,
cause crash in different areas.

The fix is to return correct status back to camera service when no camera attach
to allow the camera app/service end gracefully.

Change-Id: I4458677166ac20020637ef285b22f4300e2b9339
Signed-off-by: Chon Ming Lee <chon.ming.lee@intel.com>
Reviewed-on: https://android.intel.com/279059
Reviewed-by: Bodke, Kishore K <kishore.k.bodke@intel.com>
Tested-by: Bodke, Kishore K <kishore.k.bodke@intel.com>