[libcamerausb]: Remove memory leak from capture buffer.
Issue: BMDA-1716

There is a memory leak where capture buffer has not free.
The reason the buffer is not free is the message to free
the buffer message is submmited to the message queue with
multiple pending messages waited to be clear.
If handleMessageStopPreview message is prior to free
capture buffer message in the message queue, once that
message get executed, it removes MESSAGE_ID_RETURN_BUFFER
(for capture buffer) from message queue, the capture buffer
message will be removed and capture buffer will no longer
free and memory leak happens.

Compare to multiple cores sku, the message queue tend to
have less pending messages. As a result the message to
free the buffer can be execute right away.

Signed-off-by: Chon Ming Lee <chon.ming.lee@intel.com>

Change-Id: I04661ccc61ad244573360b9898f570bc1987ff22
Reviewed-on: https://android.intel.com/310826
Tested-by: Anuar, Nuhairi <nuhairi.anuar@intel.com>
Reviewed-by: Anuar, Nuhairi <nuhairi.anuar@intel.com>
Signed-off-by: chuckche <chuck.chen@intel.com>
Reviewed-on: https://android.intel.com/312058
Reviewed-by: sys_pccgand <sys_pccgand@intel.com>
Reviewed-by: Zhi, Yong <yong.zhi@intel.com>
Reviewed-by: Deuskar, Prafulla <prafulla.deuskar@intel.com>
Reviewed-by: Kandru, Suneel <suneel.kandru@intel.com>